Refactor mínimo para separar los formularios.
