+++ app/src/main/java/com/example/awaq1/data/FormulariosDatabase.kt
@@ -4,6 +4,8 @@ import androidx.room.Database
 import android.content.Context
 import androidx.room.Room
 import androidx.room.RoomDatabase
+import com.example.awaq1.data.formularioUno.FormularioUnoDAO
+import com.example.awaq1.data.formularioUno.FormularioUnoEntity
 
 
 @Database(entities = [FormularioUnoEntity::class], version = 1, exportSchema = false)
