+++ app/src/main/java/com/example/awaq1/view/BusquedaReportes.kt
@@ -23,7 +23,7 @@ import androidx.compose.ui.unit.dp
 import androidx.navigation.NavController
 import com.example.awaq1.MainActivity
 import com.example.awaq1.R
-import com.example.awaq1.data.FormularioUnoEntity
+import com.example.awaq1.data.formularioUno.FormularioUnoEntity
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.items
 import androidx.compose.material3.MaterialTheme
